app/requirements.txt
pika
requests
python-dotenv
Pillow
pywin32; sys_platform == "win32"
